mock_servers/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
grpcio-reflection>=1.60.0
protobuf>=4.25.0
